add update methods

+++ src/main/java/com/example/bratishka/bratishkaBackEnd/controllers/BarberController.java
@@ -4,6 +4,7 @@ import com.example.bratishka.bratishkaBackEnd.models.Barber;
 import com.example.bratishka.bratishkaBackEnd.models.Shop;
 import com.example.bratishka.bratishkaBackEnd.repositories.BarberRepository;
 import com.example.bratishka.bratishkaBackEnd.services.BarberService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,21 +19,30 @@ public class BarberController {
         this.barberService = barberService;
     }
 
-    @PostMapping("/get")
-    public ResponseEntity<List<Barber>> getBarbersByShop(@RequestBody Shop shop) {
-        List<Barber> barbers = barberService.getBarbersByShop(shop);
+    @GetMapping("/by-shop")
+    public ResponseEntity<List<Barber>> getBarbersByShop(@RequestParam long shopId) {
+        List<Barber> barbers = barberService.getBarbersByShopId(shopId);
         return ResponseEntity.ok(barbers);
     }
 
-
     @PostMapping("/add")
-    public ResponseEntity<Barber> addBarber(@RequestBody Barber barber) {
-        return ResponseEntity.ok(barberService.addBarber(barber));
+    @ResponseStatus(HttpStatus.CREATED)
+    public Barber addBarber(@RequestBody Barber barber) {
+        return barberService.saveBarber(barber);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteBarber(@PathVariable long id) {
+        if (!barberService.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         barberService.deleteBarberById(id);
         return ResponseEntity.noContent().build();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Barber> updateBarber(@PathVariable long id, @RequestBody Barber barber) {
+        barber.setId(id);
+        return ResponseEntity.ok(barberService.saveBarber(barber));
     }
 }
